Move header `bit_cast.hpp` into directory `bit` from `memory`
