+++ README.md
@@ -91,7 +91,7 @@ Procedures for each standard are provided by the following R7RS-style libraries:
 cmake -B build -DCMAKE_BUILD_TYPE=Release
 cd build
 make package
-sudo apt install build/meevax_0.5.83_amd64.deb
+sudo apt install build/meevax_0.5.84_amd64.deb
 ```
 
 or
@@ -123,9 +123,9 @@ sudo rm -rf /usr/local/share/meevax
 
 | Target Name | Description
 |-------------|-------------
-| `all`       | Build shared-library `libmeevax.0.5.83.so` and executable `meevax`
+| `all`       | Build shared-library `libmeevax.0.5.84.so` and executable `meevax`
 | `test`      | Test executable `meevax`
-| `package`   | Generate debian package `meevax_0.5.83_amd64.deb`
+| `package`   | Generate debian package `meevax_0.5.84_amd64.deb`
 | `install`   | Copy files into `/usr/local` directly
 
 ## Usage
